backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
google-generativeai>=0.3.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
numpy
